[SPARK-37823][INFRA][TESTS] Add `is-changed.py` dev script
### What changes were proposed in this pull request?

This PR aims to add `is-changed.py` dev script to reuse in CIs for various purposes in the future.
To do that, file change detection codes are moved into a new file, `dev/utils.py`.

```
$ dev/is-changed.py -h
usage: is-changed [-h] [-f] [-m MODULES]

optional arguments:
  -h, --help            show this help message and exit
  -f, --fail            Exit with 1 if there is no relevant change.
  -m MODULES, --modules MODULES
                        A comma-separated list of modules to test (default: avro,build,catalyst,core,docker-integration-tests,docs,examples,graphx,hadoop-cloud,hive,hive-
                        thriftserver,kubernetes,kvstore,launcher,mesos,mllib,mllib-local,network-common,network-shuffle,pyspark-core,pyspark-ml,pyspark-mllib,pyspark-pandas,pyspark-pandas-slow,pyspark-
                        resource,pyspark-sql,pyspark-streaming,repl,root,sketch,spark-ganglia-lgpl,sparkr,sql,sql-kafka-0-10,streaming,streaming-kafka-0-10,streaming-kinesis-asl,tags,unsafe,yarn)
```

### Why are the changes needed?

This can be used in CI to save the community resources.

### Does this PR introduce _any_ user-facing change?

No, this is a dev-only test script change.

### How was this patch tested?

This should pass the existing CIs.

The following is a manual test.
```
$ APACHE_SPARK_REF=HEAD dev/is-changed.py
false

$ APACHE_SPARK_REF=08dd010860 dev/is-changed.py
true
```

This also supports `module` arguments to check the required modules exactly.
```
$ git diff HEAD~1
diff --git a/sql/catalyst/src/main/scala/org/apache/spark/sql/internal/SQLConf.scala b/sql/catalyst/src/main/scala/org/apache/spark/sql/internal/SQLConf.scala
index 252dd5bad3..e415ee05e2 100644
--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/internal/SQLConf.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/internal/SQLConf.scala
 -1,4 +1,5
 /*
+
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
  * this work for additional information regarding copyright ownership.

$ APACHE_SPARK_REF=HEAD~1 dev/is-changed.py -m sql
true

$ APACHE_SPARK_REF=HEAD~1 dev/is-changed.py -m catalyst
true

$ APACHE_SPARK_REF=HEAD~1 dev/is-changed.py -m docs
false
```

Closes #35112 from dongjoon-hyun/SPARK-37823.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
